+++ cohttp-eio/examples/client_tls.ml
@@ -5,18 +5,24 @@ let () =
   Logs_threaded.enable ();
   Logs.Src.set_level Cohttp_eio.src (Some Debug)
 
-let null_auth ?ip:_ ~host:_ _ = Ok None     (* Warning: use a real authenticator in your code! *)
+let null_auth ?ip:_ ~host:_ _ =
+  Ok None (* Warning: use a real authenticator in your code! *)
 
 let https ~authenticator =
   let tls_config = Tls.Config.client ~authenticator () in
   fun uri raw ->
-    let host = Uri.host uri |> Option.map (fun x -> Domain_name.(host_exn (of_string_exn x))) in
+    let host =
+      Uri.host uri
+      |> Option.map (fun x -> Domain_name.(host_exn (of_string_exn x)))
+    in
     Tls_eio.client_of_flow ?host tls_config raw
 
 let () =
   Eio_main.run @@ fun env ->
   Mirage_crypto_rng_eio.run (module Mirage_crypto_rng.Fortuna) env @@ fun () ->
-  let client = Client.make ~https:(Some (https ~authenticator:null_auth)) env#net in
+  let client =
+    Client.make ~https:(Some (https ~authenticator:null_auth)) env#net
+  in
   Eio.Switch.run @@ fun sw ->
   let resp, body =
     Client.get ~sw client (Uri.of_string "https://example.com")